backend/requirements.txt
asgiref==3.9.1
click==8.3.0
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
pytokens==0.2.0
six==1.17.0
sqlparse==0.5.3
tomli==2.3.0
typing_extensions==4.15.0
